+++ app/services/socket/index.js
@@ -3,16 +3,23 @@ const ioConfig = require('./config');
 const ioState = require('./state');
 const ioEvents = require('./events');
 const ioConstants = require('./constants');
+const ioMethods = require('./methods');
+const service = require('./service');
 
 module.exports = (httpServer) => {
   const ioServer = new Server(httpServer, ioConfig);
-
-  return {
+  const ioObjects = {
     httpServer,
-    ioServer,
     ioConfig,
+    ioServer,
     ioState,
     ioEvents,
-    ioConstants
+    ioConstants,
+    ioMethods,
+  };
+
+  return {
+    ioService: () => service(ioObjects),
+    ...ioObjects,
   };
 };
