+++ app/services/socket/methods.js
@@ -22,6 +22,21 @@ const evaluateRockPaperScissors = (playerA, playerB) => {
   return null;
 };
 
+const removePlayerFromRoom = (socket) => {
+  const roomsEntries = Object.entries(state.rooms);
+  const roomIndex = roomsEntries.findIndex(([key, room]) => room[socket.id]);
+  const [roomKey, room] = rooms[roomIndex];
+
+  if (room) {
+    const [playerA, playerB] = Object.values(room);
+
+    playerA && (playerA.isPlaying = false);
+    playerB && (playerB.isPlaying = false);
+    delete state.rooms[roomKey];
+  }
+};
+
 module.exports = {
   evaluateRockPaperScissors,
+  removePlayerFromRoom
 };
